Added a macro for the main and cleaned up a bit
Trying to simplify some of the stuff.
Started a shared utils - We can definitively abstract grids / Directions

+++ aoc2024/src/days/day01.rs
@@ -49,5 +49,5 @@ pub fn solve() -> (Option<usize>, Option<usize>) {
     }
     // println!("Part 2 Result: {similarity_score}");
 
-    (Some(result as usize), Some(similarity_score as usize))
+    (Some(result as usize), Some(similarity_score))
 }
